set up component structure for image generation vs. quickstart mode
